Type-disagreement repair 3: u8 collision CLOSED, shadow class measured -- payload byte-identical
Baseline at HEAD afac8195 (not the stale gitStatus snapshot): DEFECTS.json = 949;
Stage 03's agreement test = 721 still disagreeing. Clean build 956,480 B / 97.35 %.

TWO NEW CALIBRATION BUGS IN THE INHERITED ORACLES (numbers 6 and 7 across four
passes -- assume more remain)
- The missing-argument selector compares SOURCE COMMAS against CORE REGISTERS
  READ. A long long / double / 8-byte struct argument fills TWO registers, so
  z_impl_k_sem_take(sem, -1LL) and k_mutex_lock(m, (g1_k_timeout_t){-1}) were
  flagged "missing r2, r3 = -1, -1" -- the two halves of the argument they ALREADY
  PASS. Repairing them would have INJECTED A REAL ABI DEFECT INTO WORKING CODE.
- The slot counter written to fix that then read `long long` as "a parameter named
  long". Caught by reading a diff the tool had just written, before any build.
Corrected census: the brief's 366 -> 345 sites; the carried-over 56 constant sites
-> 51.

THE GATE WAS CHANGED, DELIBERATELY AND WITH A CONTROL
Pass 2 required whole-object byte identity, which rejects DWARF-only differences.
Only ALLOCATABLE sections reach flash, so this pass gates on allocatable-byte
identity (Gate A), reporting whole-object identity (Gate B) alongside. Negative
control verified: a one-constant mutation is still caught (same size, different
bytes).

LANDED, ALL GATE-A CLEAN
- u8 TYPEDEF COLLISION CLOSED -- the highest-risk unclosed item in the project.
  Exactly one file defines u8 as 64-bit; renamed. The object grows 4 B entirely in
  .debug_str; allocatable digest unchanged.
- SHADOW CLASS MEASURED FOR THE FIRST TIME: 299 entries, 31.6 % of the ledger,
  never examined -- Stage 03's test had scored it 75 % fixed VACUOUSLY. A GCC
  conflict probe (degenerate on first run, caught by a positive control returning
  0/110, fixed, then 23/23) shows 273 OF 300 ENTRIES CONFLICT WITH THE SDK'S REAL
  DECLARATION. Adopting the SDK prototype -- harvested from GCC's own diagnostic --
  closed 37 entries / 143 declaration sites at zero flash cost.
- Canonicalisation rounds 1-2: 110 more files.

QUARANTINED WITH MEASUREMENT, NOT ARGUMENT
All 51 constant-argument sites. ZERO are free; the class costs +66 B of .text
(margin 55.6 ms, step 100.5 ms; +16 B already broke the dashboard once). 9 sites
in 8 files are size-neutral at function granularity (nm -S: 0 functions change
size) and are flagged as the cheapest first batch for a Renode-owning pass -- NOT
landed, because they change what a callee receives on the display path.

ACCEPTANCE
exit 0; nm -u 0; 0 dup globals; pin gates 0/0 both cores; thread-stack args 10/10;
verify_data 995/995; FLASH/RAM delta 0; `size -A` diff empty except .debug_*;
app_update.bin PAYLOAD BYTE-IDENTICAL (957,072 B, cmp exit 0).
Two objects differ post-strip: uname.c.obj (embedded build clock, gc'd out) and
ui_new_message_task.c.obj (a GCC switch table renumbered .CSWTCH.12 -> .14;
contents and relocations identical).
FRAMEBUFFERS OUTSTANDING, NOT CLAIMED. Renode untouched.

CORRECTIONS TO PRIOR CLAIMS
The "6 short definitions" -- already corrected to 2 by pass 2 -- is honestly 1
(slave_display_thread is a k_thread_entry_t; pass 2's refutation stands, and the
automated test cannot see it). This pass's own "64 lost calls" finding was WRONG:
per-file aliases and calls inside macro bodies; recounted to 12 callees / 29
sites, recorded as a lead rather than a defect list.

Net untouched and frozen at 225,581 B. recon/refactor/ 0 lines dirty. 673 files
modified (235 gated + 436 mirrored); `ninja -n` confirms 0 pending rebuilds, so
the mirrors are not build inputs.

Detail: recon/analysis/type_disagreement_repair_3.md

+++ recon/readable_sources/app/g1/gui_utf_Wordwrap_draw.c
@@ -56,14 +56,14 @@ typedef int i32;
 extern void log_message(unsigned long, ...);
 extern int get_device_info(void);
 extern void debug_print(unsigned long, ...);
-extern u32 get_display_atomic_state(void);
+extern unsigned long get_display_atomic_state(void);
 extern u32 get_glyph_spacing(u32,u32);
 extern i32 load_glyph_resource(u32,u32,void*,void*,void*,u32);
 extern void clear_framebuffer_region(u32,u32,u32,u32,u32,u32);
 extern void refresh_display_region(u32,u32,u32,u32,u32,u32);
 extern u16* decode_utf16_text(u32,void*);
 extern void draw_glyph_bitmap(u32,void*,u32,u32,u32,u32);
-extern u32 classify_text_character(u32);
+extern long classify_text_character(unsigned long);
 extern void copy_glyph_bitmap(void*,u32,u32,u32);
 
 typedef u32 (*fnptr_t)(u32,u32,u32,u32,u32);
