+++ recon/symbolized/app/ble_privacy_id_remove.c
@@ -37,7 +37,7 @@ extern int bt_hci_cmd_create(int,...);
 extern int bt_hci_cmd_send_sync(int,...);
 extern int bt_le_adv_legacy_invoke(int,...);
 extern int ble_conn_unref(void);
-extern int bt_conn_lookup_state_le(int, int, int);
+extern int bt_conn_lookup_state_le(unsigned long, unsigned long, unsigned long);
 extern int bt_keys_foreach_type(int,...);
 extern int net_buf_simple_add(int,...);
 extern void assert_post_action(int,...) __attribute__((noreturn));
